+++ source/r_surf.c
@@ -556,7 +556,7 @@ void R_DrawSurfaceBlock16 (void)
 		pbasesource += sourcetstep;
 		lightright += lightrightstep;
 		lightleft += lightleftstep;
-		prowdest = (unsigned short *)((long)prowdest + surfrowbytes);
+		prowdest = (unsigned short *)((intptr_t)prowdest + surfrowbytes);
 	}
 
 	prowdestbase = prowdest;
